+++ scout/server/blueprints/cases/controllers.py
@@ -4,7 +4,6 @@ import itertools
 import json
 import logging
 import os
-from base64 import b64encode
 from typing import Dict, List, Set
 
 import query_phenomizer
@@ -246,10 +245,9 @@ def bionano_case(store, institute_obj, case_obj) -> Dict:
             )
             if not fshd_loci:
                 flash(
-                    f"Sample FSHD report configured for {individual['bionano_access'].get('project')} - {individual['bionano_access'].get('sample')} but could be retrieved. Check BioNano Access server ({current_app.config.get('BIONANO_ACCESS')}) manually if the error persists.",
+                    f"Sample FSHD report configured for {individual['bionano_access'].get('project')} - {individual['bionano_access'].get('sample')} but could not be retrieved or processed. Check BioNano Access server ({current_app.config.get('BIONANO_ACCESS')}) manually if the error persists.",
                     "danger",
                 )
-
         individual["fshd_loci"] = fshd_loci
 
     data = {
@@ -381,9 +379,6 @@ def case(store, institute_obj, case_obj):
         case_obj["custom_images"] = case_obj["custom_images"].get(
             "case_images", case_obj["custom_images"].get("case", {})
         )
-        for img_section in case_obj["custom_images"].keys():
-            for img in case_obj["custom_images"][img_section]:
-                img["data"] = b64encode(img["data"]).decode("utf-8")
 
     # Limit secondary findings according to institute settings
     limit_genes = store.safe_genes_filter(institute_obj["_id"])