+++ CHANGELOG.md
@@ -3,9 +3,23 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
-## [Unreleased]
+## [unreleased]
+### Added
+- Hovertip to gene panel names with associated genes in variant view, when variant covers more than one gene
+- Tests for panel to genes
+- Download of Orphadata en_product6 and en_product4 from CLI
+### Changed
+- Allow use of projections when retrieving gene panels
+- Do not save custom images as binary data into case and variant database documents
+- Retrieve and display case and variant custom images using image's saved path
+
+## [4.74.1]
 ### Changed
 - Parse and save into database also OMIM terms not associated to genes
+### Fixed
+- BioNano API FSHD report requests are GET in Access 1.8, were POST in 1.7
+- Update more FontAwesome icons to avoid Pro icons
+- Test if files still exist before attempting to load research variants
 
 ## [4.74]
 ### Added